Refactor: Modularize CLI into domain-specific modules
