+++ src/rldk/cli/__init__.py
@@ -1,0 +1,5 @@
+"""RL Debug Kit CLI modules."""
+
+from .main import app
+
+__all__ = ["app"]
